fixed lr bug

+++ utils/parser.py
@@ -48,7 +48,7 @@ parser.add_argument('--mode',type=str,default='train',choices=['train','test','v
 parser.add_argument('--n_epochs',type=int,default=2,help= 'number of epochs to train')
 parser.add_argument('--batch_size',type=int,default=256,help= 'batch size')
 parser.add_argument('--batch_size_test',type=int,default=256,help= 'batch size for the test set')
-parser.add_argument('--lr',type=int,default=1e-02,help= 'Learning rate of the optimizer')
+parser.add_argument('--lr',type=float,default=1e-02,help= 'Learning rate of the optimizer')
 parser.add_argument('--use_scheduler',type=bool,default=True,help= 'use MultiStepLR scheduler')
 parser.add_argument('--milestones',type=list,default=[15,25,35,40],help= 'the epochs after which the learning rate is adjusted by gamma')
 parser.add_argument('--gamma',type=float,default=0.1,help= 'gamma correction to the learning rate, after reaching the milestone epochs')
